TestGraph is its own thing.
