+++ tests/classes/cells.cpp
@@ -8,7 +8,7 @@
 #include "main/dissolve.h"
 #include "math/mathFunc.h"
 #include "nodes/species.h"
-#include "tests/testing.h"
+#include "tests/testGraph.h"
 
 namespace UnitTest
 {
