+++ tests/ff/assignment-Ludwig-py5.cpp
@@ -3,7 +3,7 @@
 
 #include "data/ff/library.h"
 #include "nodes/species.h"
-#include "tests/testing.h"
+#include "tests/testGraph.h"
 #include <vector>
 
 namespace UnitTest
