+++ tests/ff/assignment-PCL2019-cations.cpp
@@ -3,7 +3,7 @@
 
 #include "data/ff/library.h"
 #include "nodes/species.h"
-#include "tests/testing.h"
+#include "tests/testGraph.h"
 
 namespace UnitTest
 {
