+++ tests/ff/benzene.cpp
@@ -5,7 +5,7 @@
 #include "kernels/force.h"
 #include "nodes/dissolve.h"
 #include "nodes/species.h"
-#include "tests/testing.h"
+#include "tests/testGraph.h"
 
 namespace UnitTest
 {
