+++ tests/ff/water-1000.cpp
@@ -4,7 +4,7 @@
 #include "kernels/energy.h"
 #include "kernels/force.h"
 #include "nodes/species.h"
-#include "tests/testing.h"
+#include "tests/testGraph.h"
 
 namespace UnitTest
 {
