+++ tests/nodes/angle.cpp
@@ -6,7 +6,7 @@
 #include "math/rangedVector3.h"
 #include "nodes/iterableGraph.h"
 #include "nodes/species.h"
-#include "tests/testing.h"
+#include "tests/testGraph.h"
 
 namespace UnitTest
 {
