+++ tests/nodes/averageMolecule.cpp
@@ -4,7 +4,7 @@
 #include "nodes/averageMolecule.h"
 #include "nodes/iterableGraph.h"
 #include "nodes/species.h"
-#include "tests/testing.h"
+#include "tests/testGraph.h"
 
 namespace UnitTest
 {
