+++ tests/nodes/dAngle.cpp
@@ -6,7 +6,7 @@
 #include "nodes/importDLPUtilsSurface.h"
 #include "nodes/iterableGraph.h"
 #include "nodes/species.h"
-#include "tests/testing.h"
+#include "tests/testGraph.h"
 
 namespace UnitTest
 {
