+++ tests/nodes/graphArgon.cpp
@@ -7,7 +7,7 @@
 #include "data/structureFactors.h"
 #include "nodes/neutronSQ.h"
 #include "nodes/sq.h"
-#include "tests/testing.h"
+#include "tests/testGraph.h"
 
 namespace UnitTest
 {
