+++ tests/nodes/histogramCN.cpp
@@ -5,7 +5,7 @@
 #include "nodes/angle.h"
 #include "nodes/iterableGraph.h"
 #include "nodes/species.h"
-#include "tests/testing.h"
+#include "tests/testGraph.h"
 
 namespace UnitTest
 {
