+++ tests/nodes/intraAngle.cpp
@@ -6,8 +6,7 @@
 #include "math/rangedVector3.h"
 #include "nodes/iterableGraph.h"
 #include "nodes/species.h"
-#include "tests/testing.h"
-#include <vector>
+#include "tests/testGraph.h"
 
 namespace UnitTest
 {
