+++ tests/nodes/neutronSQ.cpp
@@ -5,7 +5,7 @@
 #include "math/windowFunction.h"
 #include "nodes/gr.h"
 #include "nodes/sq.h"
-#include "tests/testing.h"
+#include "tests/testGraph.h"
 
 namespace UnitTest
 {
