+++ tests/nodes/qSpecies.cpp
@@ -3,7 +3,7 @@
 
 #include "nodes/qSpecies.h"
 #include "nodes/species.h"
-#include "tests/testing.h"
+#include "tests/testGraph.h"
 
 namespace UnitTest
 {
