+++ tests/nodes/sdf.cpp
@@ -5,7 +5,7 @@
 #include "nodes/importDLPUtilsPDens.h"
 #include "nodes/iterableGraph.h"
 #include "nodes/species.h"
-#include "tests/testing.h"
+#include "tests/testGraph.h"
 
 namespace UnitTest
 {
