--- tests/testingTOML.cpp
@@ -1,33 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-or-later
-// Copyright (c) 2026 Team Dissolve and contributors
-
-#include "tests/testing.h"
-
-namespace UnitTest
-{
-// Comparie TOML values with context, but without insisting on a specific ordering of fields
-void compareToml(std::string location, SerialisedValue toml, SerialisedValue toml2)
-{
-    if (toml.is_table())
-    {
-        ASSERT_TRUE(toml2.is_table()) << location;
-        for (auto &[k, v] : toml.as_table())
-        {
-            ASSERT_TRUE(toml2.contains(k)) << location << "." << k << std::endl << "Expected:" << std::endl << toml[k];
-            compareToml(std::format("{}.{}", location, k), v, toml2.at(k));
-        }
-    }
-    else if (toml.is_array())
-    {
-        auto arr = toml.as_array();
-        auto arr2 = toml2.as_array();
-        ASSERT_EQ(arr.size(), arr2.size()) << location << std::endl << "Expected" << std::endl << toml;
-        for (int i = 0; i < arr.size(); ++i)
-            compareToml(std::format("{}[{}]", location, i), arr[i], arr2[i]);
-    }
-    else
-    {
-        EXPECT_EQ(toml, toml2) << location;
-    }
-}
-}; // namespace UnitTest
